create states helper module
